Replace express-async-handler with express-async-errors
The express-async-handler package was replaced with the express-async-errors package. This change required modifying every place where the async-handler was imported and used in different functions across authController, usersController, notesController, and server files. The async-handler was removed from the functions and imported in the server file. Package.json and package-lock.json were updated to reflect this package change.