Refactor query and validation setup in notesApiSlice
The changes made in the notesApiSlice.js were related to the way we set up the query for '/notes'. Formerly a function, it was modified into a function that returns an object. The object contains the 'url' property, and the 'validateStatus' function which confirms whether the server's response is successful and there is no error.